Fixes for gcc-4.7.1
This fixes some bugs in our code that were exposed by
gcc-4.7.1 and fixes/disables some warnings.

In addition, it works around a bug in gcc-4.7.1
(apparently fixed in 4.7.2) that ends a temprary's
lifetime too soon.

+++ src/compiler/analysis/emitter.cpp
@@ -5771,8 +5771,8 @@ bool EmitterVisitor::canEmitBuiltinCall(FunctionCallPtr fn,
     return false;
   }
   if (!func->isUserFunction() && !func->needsActRec()
-      && (uint)numParams >= func->getMinParamCount()
-      && (uint)numParams <= func->getMaxParamCount()) {
+      && numParams >= func->getMinParamCount()
+      && numParams <= func->getMaxParamCount()) {
     if (func->isVariableArgument() || func->isReferenceVariableArgument()
                 || func->isMixedVariableArgument()) {
       return false;
